refactor(app): calcheck: update tip pickup animations and failure screen
