+++ app/src/components/CalibrateTipLength/TipPickUp.js
@@ -10,8 +10,8 @@ import type { CalibrateTipLengthChildProps } from './types'
 import styles from './styles.css'
 
 // TODO: put these assets in a shared location?
-import multiDemoAsset from '../CheckCalibration/videos/A1-Multi-Channel-SEQ.webm'
-import singleDemoAsset from '../CheckCalibration/videos/A1-Single-Channel-SEQ.webm'
+import multiDemoAsset from '../CheckCalibration/videos/A1_Multi_Channel_REV1.webm'
+import singleDemoAsset from '../CheckCalibration/videos/A1_Single_Channel_REV1.webm'
 
 const TIP_PICK_UP_HEADER = 'Position pipette over '
 const TIP_PICK_UP_BUTTON_TEXT = 'Pick up tip'
