+++ app/src/components/CheckCalibration/TipPickUp.js
@@ -8,10 +8,10 @@ import type { JogAxis, JogDirection, JogStep } from '../../http-api-client'
 import { getLatestLabwareDef } from '../../getLabware'
 import { JogControls } from '../JogControls'
 import styles from './styles.css'
-import multiA1DemoAsset from './videos/A1-Multi-Channel-SEQ.webm'
-import singleA1DemoAsset from './videos/A1-Single-Channel-SEQ.webm'
-import multiB1DemoAsset from './videos/B1-Multi-Channel-SEQ.webm'
-import singleB1DemoAsset from './videos/B1-Single-Channel-SEQ.webm'
+import multiA1DemoAsset from './videos/A1_Multi_Channel_REV1.webm'
+import singleA1DemoAsset from './videos/A1_Single_Channel_REV1.webm'
+import multiB1DemoAsset from './videos/B1_Multi_Channel_REV1.webm'
+import singleB1DemoAsset from './videos/B1_Single_Channel_REV1.webm'
 
 const TIP_PICK_UP_HEADER = 'Position pipette over '
 const TIP_PICK_UP_BUTTON_TEXT = 'Pick up tip'
